backend/requirements.txt
django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
pre-commit==3.3.3
psycopg==3.1.9
dj-database-url==2.0.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
redis==4.6.0
django-celery==3.1.17
django-celery-beat==2.5.0
django-celery-results==2.5.1
psycopg2==2.9.9
django-rest-multiple-models==2.1.3
requests~=2.31.0
django-colorfield==0.11.0
